Extract register banner into new component
