+++ ts/pages/governance/vote_index.tsx
@@ -1,34 +1,30 @@
 import { BigNumber } from '@0x/utils';
-import { Web3Wrapper } from '@0x/web3-wrapper';
 import { Contract, providers } from 'ethers';
 import { request, gql } from 'graphql-request';
 import * as _ from 'lodash';
 import CircularProgress from 'material-ui/CircularProgress';
 import moment from 'moment';
 import * as React from 'react';
-import { useSelector } from 'react-redux';
 import {
     useQuery,
 } from 'react-query';
-import MediaQuery  from 'react-responsive';
 import styled from 'styled-components';
 
 import { Button } from 'ts/components/button';
 import { DocumentTitle } from 'ts/components/document_title';
+import { RegisterBanner } from 'ts/components/governance/register_banner';
 import { Column, Section } from 'ts/components/newLayout';
 import { StakingPageLayout } from 'ts/components/staking/layout/staking_page_layout';
 import { Heading, Paragraph } from 'ts/components/text';
 import { Text } from 'ts/components/ui/text';
 import { Proposal, proposals as prodProposals, stagingProposals, TreasuryProposal } from 'ts/pages/governance/data';
 import { VoteIndexCard } from 'ts/pages/governance/vote_index_card';
-import { State } from 'ts/redux/reducer';
 import { colors } from 'ts/style/colors';
-import { AccountReady, TallyInterface, VotingCardType, WebsitePaths } from 'ts/types';
+import { TallyInterface, VotingCardType } from 'ts/types';
 import { ALCHEMY_API_KEY, configs, GOVERNOR_CONTRACT_ADDRESS, GOVERNANCE_THEGRAPH_ENDPOINT } from 'ts/utils/configs';
 import { constants } from 'ts/utils/constants';
 import { documentConstants } from 'ts/utils/document_meta_constants';
 import { environments } from 'ts/utils/environments';
-import { formatZrx } from 'ts/utils/format_number';
 import { OnChainProposal } from 'ts/types';
 
 const FETCH_PROPOSALS = gql`
@@ -152,10 +148,6 @@ interface Proposals {
     [index: string]: any;
 }
 
-interface TreasureContextType {
-    proposals: Proposals;
-}
-
 export const VoteIndex: React.FC<VoteIndexProps> = () => {
     const [filter, setFilter ] = React.useState<string>('all');
     const [tallys, setTallys] = React.useState<ZeipTallyMap>(undefined);
@@ -169,22 +161,6 @@ export const VoteIndex: React.FC<VoteIndexProps> = () => {
         const { proposals } = await request(GOVERNANCE_THEGRAPH_ENDPOINT, FETCH_PROPOSALS)
         return proposals;
     });
-
-    const providerState = useSelector((state: State) => state.providerState);
-    React.useEffect(() => {
-        if(providerState) {
-            const { zrxBalanceBaseUnitAmount } = providerState.account as AccountReady;
-            let zrxBalance: BigNumber | undefined;
-            if (zrxBalanceBaseUnitAmount) {
-                zrxBalance = Web3Wrapper.toUnitAmount(zrxBalanceBaseUnitAmount, constants.DECIMAL_PLACES_ZRX);
-            }
-
-            if(zrxBalance) {
-                const roundedZrxBalance = formatZrx(zrxBalance).roundedValue;
-                setZRXBalance(roundedZrxBalance);
-            }
-        }
-    }, [providerState])
 
     const abi = [
         'function quorumThreshold() public view returns (uint)',
@@ -256,35 +232,7 @@ export const VoteIndex: React.FC<VoteIndexProps> = () => {
 
     return (
         <StakingPageLayout isHome={false} title="0x Governance">
-            {
-                userZRXBalance && userZRXBalance > 0 &&
-                <RegisterBanner>
-                    <BannerImage src="/images/governance/register_banner.svg" />
-                        <MediaQuery minWidth={768}>
-                            <TextContent>
-                                    <Text noWrap={true} fontColor={colors.textDarkPrimary} Tag='h1' fontSize='28px' fontFamily='Formular'>
-                                        Register to vote with your ZRX!
-                                    </Text>
-                                    <Text noWrap={true} fontColor={colors.textDarkSecondary} fontSize='22px' fontFamily='Formular' fontWeight={300}>
-                                        Register to vote on upcoming treasury proposals
-                                    </Text>
-                            </TextContent>
-                        </MediaQuery>
-                        <MediaQuery maxWidth={768}>
-                            <TextContent>
-                                <Text noWrap={true} fontColor={colors.textDarkPrimary} Tag='h1' fontSize='22px'>
-                                    Register to vote with your ZRX!
-                                </Text>
-                                <Text noWrap={true} fontColor={colors.textDarkSecondary} fontSize='14px'>
-                                    Register to vote on upcoming treasury proposals
-                                </Text>
-                            </TextContent>
-                        </MediaQuery>
-                    <Button to={WebsitePaths.Register} color={colors.white}>
-                        Register your ZRX
-                    </Button>
-                </RegisterBanner>
-            }
+            <RegisterBanner />
             <DocumentTitle {...documentConstants.VOTE} />
             <Section isTextCentered={true} isPadded={true} padding="80px 0 80px">
                 <Column>
@@ -389,46 +337,6 @@ const LoaderWrapper = styled.div`
     display: flex;
     align-items: center;
     justify-content: center;
-`;
-
-const RegisterBanner = styled.div`
-    background-color: ${({ theme }) => theme.lightBgColor};
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 100%;
-    padding-bottom: 20px;
-
-    @media (min-width: 768px) {
-        height: 120px;
-        flex-direction: row;
-        padding-right: 50px;
-        padding-bottom: 0px;
-        width: 90%;
-        margin: auto;
-    }
-`;
-
-const BannerImage = styled.img`
-    height: 100%;
-`;
-
-const TextContent = styled.div`
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-    flex: 1;
-    justify-content: center;
-    white-space: normal;
-    text-align: center;
-    margin: 20px 0;
-
-    @media (min-width: 768px) {
-        margin: 0;
-        padding: 0 50px;
-        white-space: nowrap;
-        text-align: left;
-    }
 `;
 
 const Wrapper = styled.div`
